pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "envquest"
version = "0.0.1"
authors = [
    {name = "Medric Sonwa", email="medric49@gmail.com"}
]
description = "A collection of Reinforcement Learning algorithms to train autonomous agents in different environments."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "fire>=0.7.0",
    "gymnasium[box2d]>=1.0.0",
    "imageio[ffmpeg,pyav]>=2.35.1",
    "opencv-python>=4.10.0.84",
    "pillow>=10.4.0",
    "torch>=2.5.1",
    "tqdm>=4.67.1",
    "wandb[media]>=0.19.1",
]

[project.urls]
Homepage = "https://github.com/medric49/rlstudio"
Issues = "https://github.com/medric49/rlstudio/issues"


[dependency-groups]
dev = [
    "black>=24.8.0",
    "build>=1.2.2.post1",
    "jupyter>=1.1.1",
    "pylint>=3.2.7",
    "twine>=6.0.1",
]

[tool.setuptools]
packages = ["envquest"]


[tool.uv]
native-tls = true

[tool.black]
line-length = 120
target-version = ["py310"]
color = true

[tool.pylint.main]
disable = [
    "line-too-long",
    "missing-module-docstring",
    "missing-function-docstring",
    "missing-class-docstring",
    "duplicate-code",
    "too-few-public-methods",
    "no-member",
    "fixme",
    "logging-too-many-args",
    "no-name-in-module",
    "too-many-arguments",
    "too-many-positional-arguments",
    "too-many-locals",
    "useless-parent-delegation",
    "too-many-instance-attributes",
    "too-many-branches"
]
ignore-paths = ["./.venv"]
logging-format-style = "new"
